Adding server structure to GitHub
I didn't add server from the beginning to the GitHub, that's why I'm adding it now. For now it has 'launcher' only functions. I'm currently working on /login route and trying to resolve which approach should i use for token transfer.

For now it has working /register and /verify functions.